Move StatusHistory to separate file.
